+++ common/models/src/media/omdb-metadata.ts
@@ -30,11 +30,11 @@ export type FetchedOmdbMetadata = {
   imdbRating: string // '6.7'
   imdbVotes: string // '422,380'
   imdbID: string // 'tt0800080'
-  Type: 'movie' | 'series' // 'movie'
-  DVD: string // 'N/A'
-  BoxOffice: string // 'N/A'
-  Production: string // 'N/A'
-  Website: string // 'N/A'
+  Type: 'movie' | 'episode' // 'movie'
+  DVD?: string // 'N/A'
+  BoxOffice?: string // 'N/A'
+  Production?: string // 'N/A'
+  Website?: string // 'N/A'
   Response: 'True'
   seriesID?: string
   Season?: string