+++ services/media/src/utils/get-ffprobe-data.ts
@@ -1,7 +1,7 @@
-import { media } from '@common/models'
 import { execAsync } from '@common/service-utils'
+import { FFProbeResult } from 'ffprobe'
 
-export const getFfprobeData = async (moviePath: string): Promise<media.FfprobeMetadata> => {
+export const getFfprobeData = async (moviePath: string): Promise<FFProbeResult> => {
   const ffprobecmd = `ffprobe -v quiet -print_format json -show_format -show_streams -i "${moviePath}"`
   const ffprobeResult = await execAsync(ffprobecmd, {})
   return JSON.parse(ffprobeResult)